requirements.txt
# TERAG AI-REPS Requirements
# Core AI/ML dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.10.0
numpy>=1.24.0

# Cognitive and telemetry
chromadb>=0.4.0
pandas>=2.0.0

# Async and HTTP
aiohttp>=3.11.0
httpx>=0.27.0
asyncpg>=0.29.0

# Caching and Performance
redis>=5.2.0
hiredis>=2.2.0

# Telegram Integration
aiogram>=3.0.0
APScheduler>=3.10.0

# System Context
psutil>=5.9.0
docker>=6.0.0

# Monitoring and Metrics
prometheus-client>=0.19.0
watchdog>=3.0.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
langdetect>=1.0.9

# Self-Organizing Graph
sentence-transformers>=2.2.0

# TERAG 2.1 Benchmark & Validation
haystack-ai>=2.0.0
ragas>=0.1.2
mlflow>=2.9.0
langchain>=0.2.0
langsmith>=0.1.0

# TERAG 2.1 LangGraph Core (T.R.A.C.)
langgraph>=0.2.0
langgraph-checkpoint>=2.1.0

# TERAG 2.1 PromptOps (Phase 4)
langsmith>=0.1.0

# Language detection
langdetect>=1.0.9

# Visualization
networkx>=3.0
pyvis>=0.3.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0

# Code quality and security
bandit>=1.7.5
safety>=3.6.0
pylint>=3.0.0
flake8>=6.0.0
pip-audit>=2.7.0

# Documentation and utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Billing dependencies
pymongo>=4.6.0
stripe>=7.0.0
weasyprint>=60.0
boto3>=1.34.0

# Security dependencies
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

# ML dependencies for calibration
scikit-learn>=1.3.0
numpy>=1.24.0

# Fuzzy matching dependencies
python-Levenshtein>=0.21.0
rapidfuzz>=3.0.0  # Faster alternative to Levenshtein

# Monitoring
sentry-sdk>=2.0.0

# Demo Web Interface
streamlit>=1.28.0






















